requirements.txt
pip
minio
click
pyyaml
sklearn
tabulate
gitpython>=2.1.0